bump src/bin level to jdk 11, drop some mr jars, integrate latest dependencies
