+++ jaxb-ri/boms/bom/pom.xml
@@ -32,12 +32,12 @@
     <url>https://eclipse-ee4j.github.io/jaxb-ri/</url>
 
     <properties>
-        <xml.bind-api.version>4.0.0-RC2</xml.bind-api.version>
+        <xml.bind-api.version>4.0.0-RC3</xml.bind-api.version>
         <istack.version>4.1.0-M1</istack.version>
         <fastinfoset.version>2.1.0-M1</fastinfoset.version>
         <stax-ex.version>2.1.0-M2</stax-ex.version>
-        <activation-api.version>2.1.0-RC1</activation-api.version>
-        <angus-activation.version>1.0.0-M1</angus-activation.version>
+        <activation-api.version>2.1.0</activation-api.version>
+        <angus-activation.version>1.0.0</angus-activation.version>
     </properties>
 
     <dependencyManagement>
