+++ jaxb-ri/pom.xml
@@ -81,7 +81,7 @@
         <spotbugs.exclude/>
         <spotbugs.skip>false</spotbugs.skip>
         <spotbugs.threshold>Low</spotbugs.threshold>
-        <spotbugs.version>4.3.0</spotbugs.version>
+        <spotbugs.version>4.5.0.0</spotbugs.version>
 
         <junit.version>4.13.2</junit.version>
 
@@ -91,11 +91,9 @@
         <felix.junit4osgi>1.0.0</felix.junit4osgi>
         <felix.osgi.core>6.0.0</felix.osgi.core>
         <jmockit.version>1.34</jmockit.version>
-        <resolver.version>20050927</resolver.version>
         <compiler.version>2.4.0</compiler.version>
         <mrjar.sourceDirectory>${project.basedir}/src/main/java-mr</mrjar.sourceDirectory>
-        <base.java.level>8</base.java.level>
-        <upper.java.level>9</upper.java.level>
+        <upper.java.level>11</upper.java.level>
         <oss.disallow.snapshots>true</oss.disallow.snapshots>
         <vendor.name>Eclipse Foundation</vendor.name>
         <vendor.id>org.eclipse</vendor.id>
@@ -117,14 +115,6 @@
                 <version>${junit.version}</version>
             </dependency>
             <!-- JDK dependencies -->
-            <dependency>
-                <!-- required by com.sun.tools.xjc.Options on JDK < 9
-               (com.sun.org.apache.xml.internal.resolver.CatalogManager
-                com.sun.org.apache.xml.internal.resolver.tools.CatalogResolver) -->
-                <groupId>com.sun.org.apache.xml.internal</groupId>
-                <artifactId>resolver</artifactId>
-                <version>${resolver.version}</version>
-            </dependency>
 
             <!-- CQ #20807 -->
             <!-- contains com.sun.source.tree and com.sun.source.util JDK provided packages required by jxc (ApNavigator) -->
@@ -189,7 +179,7 @@
                     <artifactId>maven-compiler-plugin</artifactId>
                     <version>3.8.1</version>
                     <configuration>
-                        <release>${upper.java.level}</release>
+                        <release>11</release>
                     </configuration>
                 </plugin>
                 <plugin>
@@ -203,6 +193,9 @@
                     <version>3.2.1</version>
                     <configuration>
                         <archive>
+                            <manifest>
+                                <addDefaultEntries>false</addDefaultEntries>
+                            </manifest>
                             <manifestEntries>
                                 <Specification-Title>Jakarta XML Binding</Specification-Title>
                                 <Specification-Version>${jaxb-api.majorVersion}.${jaxb-api.minorVersion}</Specification-Version>
@@ -225,6 +218,9 @@
                     <version>3.3.0</version>
                     <configuration>
                         <archive>
+                            <manifest>
+                                <addDefaultEntries>false</addDefaultEntries>
+                            </manifest>
                             <manifestEntries>
                                 <Specification-Title>Jakarta XML Binding</Specification-Title>
                                 <Specification-Version>${jaxb-api.majorVersion}.${jaxb-api.minorVersion}</Specification-Version>
@@ -463,32 +459,14 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>default-compile</id>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                        <configuration>
-                            <compilerArgument>-Xlint:all</compilerArgument>
-                            <!--<XDignore.symbol.file/>-->
-                            <showDeprecation>true</showDeprecation>
-                            <showWarnings>true</showWarnings>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>base-compile</id>
-                        <goals>
-                            <goal>compile</goal>
-                        </goals>
-                        <configuration>
-                            <release>${base.java.level}</release>
-                            <excludes>
-                                <exclude>module-info.java</exclude>
-                            </excludes>
-                        </configuration>
-                    </execution>
-                </executions>
+                <configuration>
+                    <release>11</release>
+                    <testRelease>11</testRelease>
+                    <compilerArgument>-Xlint:all</compilerArgument>
+                    <!--<XDignore.symbol.file/>-->
+                    <showDeprecation>true</showDeprecation>
+                    <showWarnings>true</showWarnings>
+                </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -648,14 +626,6 @@
                                 <!-- configuration should be removed once this bug https://github.com/jacoco/jacoco/issues/407 is fixed -->
                                 <configuration>
                                     <excludes>
-                                        <!-- core -->
-                                        <exclude>org/glassfish/jaxb/core/StackHelper.class</exclude>
-                                        <!-- impl -->
-                                        <exclude>org/glassfish/jaxb/runtime/v2/MUtils.class</exclude>
-                                        <!-- xjc -->
-                                        <exclude>com/sun/tools/xjc/CatalogUtil.class</exclude>
-                                        <!-- jxc -->
-                                        <exclude>com/sun/tools/jxc/SchemaGenTask.class</exclude>
                                         <!-- istack-runtime -->
                                         <exclude>com/sun/istack/logging/StackHelper.class</exclude>
                                         <!-- istack-tools -->
